Pinch Layer Plugin adapted to work with KineticJS 4.7.4
I changed this plugin to work with KineticJS 4.7.4. I didn't change much. I just changed "Kinetic.Plugins.PinchLayer" to "Kinetic.PinchLayer" and I changed line 19 to "self.dStage = self.getParent().getContent();" (as .getDOM() does not exist in the stage object in KineticJS 4.7.4). I also commented out line 109 to 114 as I do not know what the new version of the transitionTo function is called (if there is one). 

It works perfectly for me and it solves a problem I've had for days now. Thanks for this fantastic plugin.